+++ lib/ui/screens/res/custom_colors.dart
@@ -2,9 +2,11 @@ import 'package:flutter/material.dart';
 
 /// Класс в котором хранятся все катосмные цвета, которые используются в приложении
 class CustomColors {
+  /// Набор цветов для светлой темы
   static var lmBackgroundColor = const Color.fromRGBO(245, 245, 245, 1);
   static var lmSightCardBackgroundColor = const Color.fromRGBO(245, 245, 245, 1);
 
+  /// Набор цветов для тёмной темы
   static var dmBackgroundColor = const Color(0xFF2E2E2E);
   static var dmSightCardBackgroundColor = const Color.fromRGBO(26, 26, 32, 1);
 }
