+++ lib/ui/screens/res/custom_text_styles.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/cupertino.dart';
+
+const largeTitle32Bold = TextStyle(
+  height: 1.12,
+  fontSize: 32,
+  fontWeight: FontWeight.w700,
+);
